# docs/static/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for AltMount — a WebDAV/NFS server backed by NZB/Usenet, with queue management, file health monitoring, and media streaming.",
        "title": "AltMount API",
        "contact": {
            "name": "AltMount GitHub Issues",
            "url": "https://github.com/javi11/altmount/issues"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/javi11/altmount/blob/master/LICENSE"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/system/browse": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lists files and directories at the given absolute path on the server filesystem",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Browse filesystem",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Absolute directory path to browse (defaults to working directory)",
                        "name": "path",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/system/cleanup": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes old queue items and health records based on configurable age thresholds",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "System cleanup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Remove queue items older than this timestamp",
                        "name": "queue_older_than",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Remove health records older than this timestamp",
                        "name": "health_older_than",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Simulate cleanup without actually deleting",
                        "name": "dry_run",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.SystemCleanupResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/api/system/restart": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Initiates a server restart; the response is returned before the restart occurs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Restart the server",
                "parameters": [
                    {
                        "description": "Restart options",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/api.SystemRestartRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.SystemRestartResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/system/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns combined queue, health, and system statistics",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Get system statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.SystemStatsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/system/stats/reset": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Resets NNTP pool metrics, import history, and optionally queue items",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Reset system statistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reset statistics for last N duration (e.g. 24h)",
                        "name": "duration",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Reset peak speed metrics",
                        "name": "reset_peak",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Reset total bytes/articles counters",
                        "name": "reset_totals",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Reset import history",
                        "name": "reset_history",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Clear completed and failed queue items",
                        "name": "reset_queue",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "api.APIMeta": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "api.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/api.APIError"
                },
                "meta": {
                    "$ref": "#/definitions/api.APIMeta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "api.HealthStatsResponse": {
            "type": "object",
            "properties": {
                "checking": {
                    "type": "integer"
                },
                "corrupted": {
                    "type": "integer"
                },
                "healthy": {
                    "type": "integer"
                },
                "pending": {
                    "type": "integer"
                },
                "repair_triggered": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "api.SystemCleanupResponse": {
            "type": "object",
            "properties": {
                "dry_run": {
                    "type": "boolean"
                },
                "health_records_removed": {
                    "type": "integer"
                },
                "queue_items_removed": {
                    "type": "integer"
                }
            }
        },
...
        "api.SystemRestartRequest": {
            "type": "object",
            "properties": {
                "force": {
                    "description": "Force restart even if unsafe",
                    "type": "boolean"
                }
            }
        },
        "api.SystemRestartResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "api.SystemStatsResponse": {
            "type": "object",
            "properties": {
                "health": {
                    "$ref": "#/definitions/api.HealthStatsResponse"
                },
                "queue": {
                    "$ref": "#/definitions/api.QueueStatsResponse"
                },
                "system": {
                    "$ref": "#/definitions/api.SystemInfoResponse"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API key for programmatic access (available from GET /api/config)",
            "type": "apiKey",
            "name": "apikey",
            "in": "query"
        },
        "BearerAuth": {
            "description": "JWT token obtained from POST /api/auth/login. Format: \"Bearer \u003ctoken\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}